refactor(brett): ws-handler.ts → ws-connection (Lifecycle) + ws-handler (Dispatcher) (T000661)
